improve gap size calculation

+++ src/device-tracker.h
@@ -20,7 +20,7 @@ typedef struct _device_tracker_t device_tracker_t;
 
 extern device_tracker_t* device_tracker_new(zlist_t* known_devices, zsock_t* sub_socket);
 extern void device_tracker_destroy(device_tracker_t** tracker);
-extern int64_t device_tracker_calculate_gap(device_tracker_t* tracker, msg_meta_t* meta, const char* pub_spec);
+extern size_t device_tracker_calculate_gap(device_tracker_t* tracker, msg_meta_t* meta, const char* pub_spec);
 extern void device_tracker_reconnect_stale_devices(device_tracker_t* tracker);
 
 extern bool log_gaps;
